Confirmation alert box on desktop before remove other device

+++ browser/resources/settings/brave_sync_page/brave_sync_configure.js
@@ -118,6 +118,12 @@ Polymer({
   },
 
   onDeleteDevice_: async function(e) {
+    const messageText = this.i18n('braveSyncDeleteDeviceConfirmation')
+    const shouldDeleteDevice = confirm(messageText)
+    if (!shouldDeleteDevice) {
+      return
+    }
+
     const deviceGuid = e.currentTarget.arg
     await this.browserProxy_.deleteDevice(deviceGuid);
   }
